creacion de admin y validaciones de permisos

+++ admin/consult_roles.php
@@ -18,6 +18,12 @@ $role = isset($_GET['rol']) ? $_GET['rol'] : '';
 $permission = isset($_GET['permiso']) ? $_GET['permiso'] : '';
 $id_role = isset($_GET['id_rol']) ? $_GET['id_rol'] : '';
 
+$create = isset($_SESSION['Crear ROL']) ? $_SESSION['Crear ROL'] : 0 ;
+$update = isset($_SESSION['Actualizar ROL']) ? $_SESSION['Actualizar ROL'] : 0 ;
+$delete = isset($_SESSION['Eliminar ROL']) ? $_SESSION['Eliminar ROL'] : 0 ;
+$deleteRP = isset($_SESSION['Eliminar ROL_PERMISO']) ? $_SESSION['Eliminar ROL_PERMISO'] : 0 ;
+$createRP = isset($_SESSION['Crear ROL_PERMISO']) ? $_SESSION['Crear ROL_PERMISO'] : 0 ;
+
 $answer = '';
 
 if ($option !== '') {
@@ -32,13 +38,13 @@ if ($answer == '') {
   try {
     $connect = new AdminSQL();
 
-    if($option === 'update') {
+    if(($option === 'update') && ($update === 1)) {
       $answer = $connect->updateRole($id_role, $role);
-    }else if($option === 'delete'){
+    }else if(($option === 'delete') && ($delete === 1)){
       $answer = $connect->deleteRole($id_role);
     }else if($option === 'select'){
       $answer = $connect->getRoles();
-    } else if($option === 'create') {
+    } else if(($option === 'create') && ($create === 1)) {
       $answer = $connect->createRole($role);
     } else if($option === 'permisos') {
       $answer = $connect->getPermissionsForRole($role);
@@ -46,9 +52,9 @@ if ($answer == '') {
       $answer = $connect->getAllPermissions();
     } else if($option === 'all-roles') {
       $answer = $connect->getRoles();
-    } else if ($option === 'associate') {
+    } else if (($option === 'associate') && ($createRP === 1)) {
       $answer = $connect->associateRoleAndPermission($role, $permission);
-    } else if ($option === 'unlink') {
+    } else if (($option === 'unlink') && ($deleteRP === 1)) {
       $answer = $connect->unlinkRoleAndPermission($role, $permission);
     }
